fix(core): move more stuff to shared
metriport/metriport-internal#1040

Ref: ENG-257
