+++ packages/core/src/command/hl7v2-subscriptions/hl7v2-to-fhir-conversion/shared.ts
@@ -166,40 +166,3 @@ export function createFileKeyHl7Message({
 export function formatDateToHl7(date: Date): string {
   return buildDayjs(date).format("YYYYMMDDHHmmss");
 }
-
-export function hl7ToIso8601(hl7DateString: string): string {
-  if (!hl7DateString || hl7DateString.length !== 14) {
-    throw new MetriportError("Invalid HL7 date string format: expected YYYYMMDDHHMMSS");
-  }
-
-  if (!/^\d{14}$/.test(hl7DateString)) {
-    throw new MetriportError("Invalid HL7 date string: must contain only digits");
-  }
-
-  // Parse YYYYMMDDHHMMSS format
-  const year = hl7DateString.substring(0, 4);
-  const month = hl7DateString.substring(4, 6);
-  const day = hl7DateString.substring(6, 8);
-  const hour = hl7DateString.substring(8, 10);
-  const minute = hl7DateString.substring(10, 12);
-  const second = hl7DateString.substring(12, 14);
-
-  // Validate date components
-  const monthNum = parseInt(month);
-  const dayNum = parseInt(day);
-  const hourNum = parseInt(hour);
-  const minuteNum = parseInt(minute);
-  const secondNum = parseInt(second);
-
-  if (monthNum < 1 || monthNum > 12) {
-    throw new MetriportError("Invalid month in HL7 date string");
-  }
-  if (dayNum < 1 || dayNum > 31) {
-    throw new MetriportError("Invalid day in HL7 date string");
-  }
-  if (hourNum > 23 || minuteNum > 59 || secondNum > 59) {
-    throw new MetriportError("Invalid time in HL7 date string");
-  }
-
-  return `${year}-${month}-${day}T${hour}:${minute}:${second}Z`;
-}
